Changed data types to Integer

+++ src/main/java/com/example/nd_medicine/controller/AppointmentController.java
@@ -67,13 +67,13 @@ public class AppointmentController {
 
     //all appointments for a particular patient -- useful in patient profile
     @GetMapping(value = "/patient/{patientId}")
-    public List<Appointment> viewAppointmentOfPatient(@PathVariable Long patientId){
+    public List<Appointment> viewAppointmentOfPatient(@PathVariable Integer patientId){
         return appointmentService.findByPatientId(patientId);
     }
 
     //all appointments for a particular patient -- useful in doc profile
     @GetMapping(value = "/doctor/{doctorId}")
-    public List<Appointment> viewAppointmentOfDoctor(@PathVariable Long doctorId){
+    public List<Appointment> viewAppointmentOfDoctor(@PathVariable Integer doctorId){
         return appointmentService.findByDoctorId(doctorId);
     }
